更新car_config.h以调整角度打印计数和手动设置初始角度的宏定义；优化Task.c和Test.c中的初始角度处理逻辑和调试打印计数器类型。

+++ control/CONTROL/Test.c
@@ -131,7 +131,7 @@ void turnToAbsoluteAngle_TEST_Handler(void)
  */
 void print_angle_Handle(void)
 {
-    static uint16_t debug_print_counter = 0; // 调试信息发送计数器
+    static uint32_t debug_print_counter = 0; // 调试信息发送计数器
     bool can_print_debug = false;
     
     if (++debug_print_counter >= ANGLE_PRINT_COUNT) {
